Extract common attributes and behaviors to from Address, Email and Phone classes to Contact class
